Debugging
After some extensive testing, I've been able to discover that the error seems to be from Spring > out. The 401 on postman is happening not because of SQL, but because something in Spring isn't right. At least, that is what I've been able to get back from testing...going to try to get some kind of auth involved to see if it's because of Spring Web Security dependency, since the methods appear to be correct but it's a 401. If we can get away from the 401, we made progress.